==> examples/experimental/2.0-add_(target)_variables_and_make_subset_ready_for_synthetic_generation.ipynb <==
import pandas as pd
import numpy as np

# Make standard dataset as how it has to be as input for synthetic data

df = pd.read_csv('C:\Users\bde2002.53381\Desktop\master-thesis\experiments\data\input\df2_subset.csv')

### Add target variable to dataset

# Years of survival for patients that died
df["pat_overlijden_dtm"] = pd.to_datetime(df['pat_overlijden_dtm'])
df['tum_incidentie_dtm'] = pd.to_datetime(df['tum_incidentie_dtm'])
df['days_of_survival'] = (df['pat_overlijden_dtm'] - df['tum_incidentie_dtm']).dt.days

# Add days of survival for patients that were alive at 31-12-2018
# Note that this is only useful to see which patients survived for X years yes/no, it is not their final years of survival
df.loc[df.vitstat == 0, 'days_of_survival'] = (pd.to_datetime("2019-01-01") - df['tum_incidentie_dtm']).dt.days

# Add binary target variable, 1460 days is four years which leads to approx even label division (66688 (True) vs 71517 (False))
# If choosing for 5 years the division will be 55862 (True) vs 76257 (False)
# Patients alive at 31-12-2018 but did not pass the fours years yet, are deleted from the dataset
df['4_yos'] = False
df.loc[df.days_of_survival > 1459, '4_yos'] = True

df = df.drop(df[(df.vitstat == 0) & (df['4_yos'] == False)].index)
df = df.drop(columns=['days_of_survival']);
df["4_yos"] = df["4_yos"].replace([True, False],[1, 0])

# Same but for 5 years of survival (makes more sense in this domain) 42% survived
df['5_yos'] = False
df.loc[df.days_of_survival > 1824, '5_yos'] = True

df = df.drop(df[(df.vitstat == 0) & (df['5_yos'] == False)].index)
df = df.drop(columns=['days_of_survival']);
df["5_yos"] = df["5_yos"].replace([True, False],[1, 0])

### Compute and remove variables

# Add age at diagnosis
df['pat_geboorte_dtm'] = pd.to_datetime(df['pat_geboorte_dtm'])
df['diagnosis_age'] = (df['tum_incidentie_dtm'] - df['pat_geboorte_dtm']).dt.days

# Make integer instead of float
df['tum_topo_sublokalisatie_code'] = df['tum_topo_sublokalisatie_code'].astype('int64')

# Remove date of birth, incidence date, date of death, registration number, vital status
df = df.drop(columns=['rn', 'pat_geboorte_dtm', 'tum_incidentie_dtm', 'pat_overlijden_dtm', 'vitstat']);

# Set category "unknown" to missing value so variables become numerical and ordinal
df["tum_differentiatiegraad_code"] = df["tum_differentiatiegraad_code"].replace([9],[np.nan])
df["tum_lymfklieren_positief_atl"] = df["tum_lymfklieren_positief_atl"].replace([98,999],[np.nan, np.nan])

# Set target to end
df = df[[col for col in df.columns if col != '5_yos'] + ['5_yos']]

## Now dataset is ready to generate synthetic data

# Save as dataset now that it has a target variable
df.to_csv("C:\Users\bde2002.53381\Desktop\master-thesis\experiments\data\df2_target.csv", index=False) # Also save in Data folder

# Read dataset for background information and descriptives (removed due to data confidentiality)

df = pd.read_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\df2_target.csv')

## Correlation matrix of dataset

## Feature distributions
