==> examples/experimental/3.0-split_dataset_up_into_folds_for_multiple_Xori_and_Xval_datasets.ipynb <==
import pandas as pd
from sklearn.model_selection import StratifiedKFold
import numpy as np

# Load final dataset with target to generate synthetic data with
df = pd.read_csv("C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\df2_target.csv")

X = df.iloc[:,:-1]
y = df.iloc[:,-1]

# Split into 5 stratified folds, so 5 "different" original sets with corresponding holdout validation sets are created
skf = StratifiedKFold(n_splits=5, random_state=8, shuffle=True)
print(skf)

# Save indices per split so separate dataframes can be saved later
train_idx = []
val_idx = []
for train_index, test_index in skf.split(X, y):
    train_idx.append(train_index)
    val_idx.append(test_index)

# Create the 5 dataframes including corresponding holdout validation set
df_ori_1 = df.loc[train_idx[0]]
df_val_1 = df.loc[val_idx[0]]

df_ori_2 = df.loc[train_idx[1]]
df_val_2 = df.loc[val_idx[1]]

df_ori_3 = df.loc[train_idx[2]]
df_val_3 = df.loc[val_idx[2]]

df_ori_4 = df.loc[train_idx[3]]
df_val_4 = df.loc[val_idx[3]]

df_ori_5 = df.loc[train_idx[4]]
df_val_5 = df.loc[val_idx[4]]

# Save the dataframes
all_ori_dfs = [df_ori_1, df_ori_2, df_ori_3, df_ori_4, df_ori_5]
all_val_dfs = [df_val_1, df_val_2, df_val_3, df_val_4, df_val_5]

i = 1
for df in all_ori_dfs:
    df.to_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\df_ori_{}.csv'.format(i), index=False)
    i += 1
    
i = 1
for df in all_val_dfs:
    df.to_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\df_val_{}.csv'.format(i), index=False)
    i += 1
