==> examples/experimental/6.0-data_preparation_for_QMs-P.ipynb <==
import pandas as pd
import numpy as np

## Data preparation
#### 1. Handle and impute missing values
#### 2. Normalize/standardize numerical values --> moved to quality measure function
#### 3. Encode categorical data

def impute_missing_values():
    missing_vals = df.isnull().sum()

    for i, val in enumerate(missing_vals):
        if missing_vals[i] > 0:
            df.iloc[:,i] = df.iloc[:,i].replace([np.nan],[df.iloc[:,i].mode()])

DPalgs = ['ms','pb']
epsilon_names = ['10','1','0_1','0_01','0_001']

# Define nominal categorical variables
nominal_cols = ['tum_topo_sublokalisatie_code', 'stadium', "binned_diagnosis_age", "binned_tum_lymphnodes_pos"]
prefix = ['subloc', 'stage', 'diagnosis_age', 'lymph_pos']

for alg in DPalgs:
    for i in range(1,6):
        for epsilon in epsilon_names:
            for versie in range(1,6):
                # Load dataframe
                df = pd.read_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\{}_df_syn_{}_e{}_v{}.csv'.format(alg, i, epsilon, versie))
                
                # Replace NaN values with 'unknown' category for lymph nodes
                df["binned_tum_lymphnodes_pos"] = df["binned_tum_lymphnodes_pos"].replace([np.nan],'unknown')
                
                # Impute remaining missing values in dataset with the mode
                impute_missing_values()
                
                # Encode the categorical variables
                df = pd.get_dummies(df, prefix=prefix, columns=nominal_cols)
                
                # Set target variable at the end
                df = df[[col for col in df.columns if col != '5_yos'] + ['5_yos']]
                
                # Save dataframe
                df.to_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\data_prepared_for_predictive_metrics\\{}\\{}_df_syn_{}_e{}_v{}_p.csv'.format(alg, alg, i, epsilon, versie), index=False)

# Also for original and validation datasets
sort_dfs = ['ori', 'val']

# Define nominal categorical variables
nominal_cols = ['tum_topo_sublokalisatie_code', 'stadium', "binned_diagnosis_age", "binned_tum_lymphnodes_pos"]
prefix = ['subloc', 'stage', 'diagnosis_age', 'lymph_pos']

for sort in sort_dfs:
    for i in range(1,6):
        df = pd.read_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\df_{}_{}_binned.csv'.format(sort, i))

        # Replace NaN values with 'unknown' category for lymph nodes
        df["binned_tum_lymphnodes_pos"] = df["binned_tum_lymphnodes_pos"].replace([np.nan],'unknown')

        # Impute remaining missing values in dataset with the mode
        impute_missing_values()

        # Encode the categorical variables
        df = pd.get_dummies(df, prefix=prefix, columns=nominal_cols)

        # Set target variable at the end
        df = df[[col for col in df.columns if col != '5_yos'] + ['5_yos']]

        # Save dataframe
        df.to_csv('C:\\Users\\bde2002.53381\\Desktop\\master-thesis\\experiments\\data\\data_prepared_for_predictive_metricsdf_{}_{}_binned_p.csv'.format(sort, i), index=False)
